fix: Checking for python version to ignore test do not work for syntax errors, splitting to a new file and ignoring import
